chore: some cleanup and organization
- remove commented out code from BaseSelect
- pull out cart helpers into their own cart.ts file
