+++ app/src/utils/utils.ts
@@ -4,9 +4,7 @@
 import router from 'src/router/index';
 import { RouteLocationRaw } from 'vue-router';
 import { BigNumber, BigNumberish, Contract, ContractTransaction, isAddress } from 'src/utils/ethers';
-import { SUPPORTED_TOKENS_MAPPING } from 'src/utils/constants';
-import { CartItem, CartItemOptions } from 'src/types';
-import { Donation, Grant, GrantRound, SwapSummary } from '@dgrants/types';
+import { GrantRound } from '@dgrants/types';
 
 // --- Formatters ---
 // Returns an address with the following format: 0x1234...abcd
@@ -42,68 +40,6 @@ export function isValidAddress(val: string | undefined) {
   return val && isAddress(val);
 }
 
-// --- Grants + Cart ---
-const CART_KEY = 'cart';
-const DEFAULT_CONTRIBUTION_TOKEN_ADDRESS = '0x6B175474E89094C44Da98b954EedeAC495271d0F'; // DAI
-const DEFAULT_CONTRIBUTION_AMOUNT = 5; // this is converted to a parsed BigNumber at checkout
-
-// Loads cart data
-export function loadCart(): CartItemOptions[] {
-  // Return empty array if nothing found
-  const rawCart = localStorage.getItem(CART_KEY);
-  if (!rawCart) return [];
-
-  // Parse the data. If the data is an array, return it,
-  const cart = JSON.parse(rawCart);
-  if (Array.isArray(cart)) return cart;
-
-  // Otherwise clear the localStorage key and return empty array
-  localStorage.removeItem(CART_KEY);
-  return [];
-}
-
-// Adds a grant to the cart
-export function addToCart(grant: Grant | null | undefined): CartItemOptions[] {
-  if (!grant) return []; // null and undefined input types are to avoid lint errors when calling this from a template
-  // If this grant is already in the cart, do nothing
-  const cart = loadCart();
-  if (cart.map((grant) => grant.grantId).includes(grant.id.toString())) return cart;
-
-  // Otherwise, add it to the cart and update localStorage
-  cart.push({
-    grantId: grant.id.toString(),
-    contributionTokenAddress: DEFAULT_CONTRIBUTION_TOKEN_ADDRESS,
-    contributionAmount: DEFAULT_CONTRIBUTION_AMOUNT,
-  });
-  localStorage.setItem(CART_KEY, JSON.stringify(cart));
-  return cart;
-}
-
-// Removes a grant from the cart
-export function removeFromCart(grantId: BigNumberish): CartItemOptions[] {
-  const cart = loadCart();
-  const newCart = cart.filter((grant) => grant.grantId !== BigNumber.from(grantId).toString());
-  localStorage.setItem(CART_KEY, JSON.stringify(newCart));
-  return newCart;
-}
-
-// Clears the cart
-export function clearCart(): CartItemOptions[] {
-  localStorage.removeItem(CART_KEY);
-  return [];
-}
-
-// Sets the full cart
-export function setCart(cart: CartItemOptions[]) {
-  localStorage.setItem(CART_KEY, JSON.stringify(cart));
-}
-
-// Check against the grantRounds status for a match
-export function hasStatus(status: string) {
-  // returns a fn (currying the given status)
-  return (round: GrantRound) => round.status === status;
-}
-
 // --- Tokens ---
 // Check for approved allowance
 export async function checkAllowance(token: Contract, ownerAddress: string | undefined, spenderAddress: string) {
@@ -120,28 +56,10 @@ export async function getApproval(token: Contract, address: string, amount: BigN
 }
 
 // --- Other ---
-// Convert a cart into an array of objects summarizing the cart info, with human-readable values
-export function getCartSummary(cart: CartItem[]): Record<keyof typeof SUPPORTED_TOKENS_MAPPING, number> {
-  const output: Record<keyof typeof SUPPORTED_TOKENS_MAPPING, number> = {};
-  for (const item of cart) {
-    const tokenAddress = item.contributionToken.address;
-    if (tokenAddress in output) output[tokenAddress] += item.contributionAmount;
-    else output[tokenAddress] = item.contributionAmount;
-  }
-  return output;
-}
-
-// Takes an array of cart items and returns inputs needed for the GrantRoundManager.donate() method
-export function formatDonateInputs(cart: CartItem[]): { swaps: SwapSummary[]; donations: Donation[] } {
-  const swaps: SwapSummary[] = [];
-  const donations: Donation[] = [];
-
-  for (const item of cart) {
-    // TODO use helper methods from @dgrants/utils
-    item;
-  }
-
-  return { swaps, donations };
+// Check against the grantRounds status for a match
+export function hasStatus(status: string) {
+  // returns a fn (currying the given status)
+  return (round: GrantRound) => round.status === status;
 }
 
 // Navigates to the specified page and pushes a new entry into the history stack
